Rename sensor-service to sensor
