+++ services/sensor/docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.4"
+version: "3.7"
 services:
   postgres:
     image: postgres
