+++ services/sensor/go.mod
@@ -1,0 +1,20 @@
+module github.com/moorara/microservices-demo/services/sensor
+
+go 1.13
+
+require (
+	github.com/DATA-DOG/go-sqlmock v1.3.3
+	github.com/codahale/hdrhistogram v0.0.0-20161010025455-3a0bb77429bd // indirect
+	github.com/go-kit/kit v0.9.0
+	github.com/google/uuid v1.1.1
+	github.com/gorilla/mux v1.7.3
+	github.com/lib/pq v1.2.0
+	github.com/moorara/konfig v0.3.2
+	github.com/opentracing/opentracing-go v1.1.0
+	github.com/prometheus/client_golang v1.2.1
+	github.com/stretchr/testify v1.4.0
+	github.com/uber-go/atomic v1.4.0 // indirect
+	github.com/uber/jaeger-client-go v2.19.0+incompatible
+	github.com/uber/jaeger-lib v2.2.0+incompatible
+	go.uber.org/atomic v1.4.0 // indirect
+)
