+++ services/sensor/init-postgres.sql
