+++ services/sensor/scripts/build.sh
