+++ services/sensor/scripts/test-component.sh
