+++ services/sensor/test/component/component_test.go
