+++ services/sensor/util/tracer_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/go-kit/kit/log"
-	"github.com/moorara/microservices-demo/services/sensor-service/config"
+	"github.com/moorara/microservices-demo/services/sensor/config"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/stretchr/testify/assert"
 )
